% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{addModelId}
\alias{addModelId}
\title{Add model_id and model overview in model_factory.model_overview table}
\usage{
addModelId(model_id, model_description, score_id_type)
}
\arguments{
\item{model_id}{The current model ID}

\item{model_description}{Model description}

\item{score_id_type}{Type of score_id (loc_id,CKR, account etc)}
}
\value{
The result of taQuery to update the model_factory.model_overview table
}
\description{
Add model_id and model overview in model_factory.model_overview table
}
\examples{
addModelId('titanic_training','Training on Titanic data','passengerid')
}

