% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{closeSession}
\alias{closeSession}
\title{Close the current session and update the end_time of run in model_factory.run_history table for current session session with the current timestamp}
\usage{
closeSession()
}
\value{
The result of taQuery to update the model_factory.run_history table
}
\description{
Close the current session and update the end_time of run in model_factory.run_history table for current session session with the current timestamp
}
\examples{
closeSession()
}

