% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{deleteModelId}
\alias{deleteModelId}
\title{Remove model_id and model overview from model_factory.model_overview table}
\usage{
deleteModelId(model_id)
}
\arguments{
\item{model_id}{The vector of model id's}
}
\value{
The result of taQuery to update the model_factory.model_overview table
}
\description{
Remove model_id and model overview from model_factory.model_overview table
}
\examples{
deleteModelId(c('titanic_training'))
}

