% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{deleteSession}
\alias{deleteSession}
\title{Remove all data associated with a session}
\usage{
deleteSession(session_id)
}
\arguments{
\item{session_id}{The vector of session id's}
}
\value{
True if successful
}
\description{
Remove all data associated with a session
}
\examples{
deleteSession(c('cvpm_t9_vecht499_20160627_949068'))
}

