% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{getConfMatrix}
\alias{getConfMatrix}
\title{Get Confusion matrix for a model of Classification type (0/1 as output) with a given treshold and threshold_type}
\usage{
getConfMatrix(scores, labels, threshold_value, threshold_type)
}
\arguments{
\item{scores}{Predicted probability based on the test set}

\item{labels}{Actual labels based on the test set}

\item{threshold_value}{threshold_value to choose}

\item{threshold_type}{Type of threshold to choose ("population" or "probability")}
}
\value{
Confusion matrix based on given scores, labels, threshold_value and threshold type
}
\description{
Get Confusion matrix for a model of Classification type (0/1 as output) with a given treshold and threshold_type
}
\examples{
train <- ta.data.frame("select * from model_factory.titanic_train",sourceType="query")
test <- ta.data.frame("select * from model_factory.titanic_test",sourceType="query")

formula <- "survived ~ sex + pclass + age + family + embarked + title"
ta_forest <- ta.forest(formula= formula, data = train,type = "classification", ntree = 100,mtry=4)
ta_forest_prediction <- ta.forest.predict(object = ta_forest, newdata=test, id.col="passengerid")

forest_output <- ta_forest_prediction$retval
forest_output$pred <- ifelse(forest_output$prediction=='Survived',as.numeric(forest_output$confidence_upper),
                             1-as.numeric(forest_output$confidence_upper))

test <- ta.pull(test)
getConfMatrix(forest_output$pred,test$survived_int,0.5,"probability")
getConfMatrix(forest_output$pred,test$survived_int,0.5,"population")
}

