% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{getLiftChart}
\alias{getLiftChart}
\title{Get liftchart table for a model of Classification type (0/1 as output)}
\usage{
getLiftChart(scores, labels)
}
\arguments{
\item{scores}{Predicted probability based on the test set}

\item{labels}{Actual labels based on the test set}
}
\value{
Liftchart table based on given scores and labels
}
\description{
Get liftchart table for a model of Classification type (0/1 as output)
}
\examples{
train <- ta.data.frame("select * from model_factory.titanic_train",sourceType="query")
test <- ta.data.frame("select * from model_factory.titanic_test",sourceType="query")

formula <- "survived ~ sex + pclass + age + family + embarked + title"
ta_forest <- ta.forest(formula= formula, data = train,type = "classification", ntree = 100,mtry=4)
ta_forest_prediction <- ta.forest.predict(object = ta_forest, newdata=test, id.col="passengerid")

forest_output <- ta_forest_prediction$retval
forest_output$pred <- ifelse(forest_output$prediction=='Survived',as.numeric(forest_output$confidence_upper),
                             1-as.numeric(forest_output$confidence_upper))

test <- ta.pull(test)
getLiftChart(forest_output$pred,test$survived_int)
}

