% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{getSessionId}
\alias{getSessionId}
\title{Get a new random session_id and store the session in Aster model_factory.run_history table}
\usage{
getSessionId(model_id)
}
\arguments{
\item{model_id}{The current model ID}
}
\value{
A new session ID
}
\description{
Get a new random session_id and store the session in Aster model_factory.run_history table
}
\examples{
model_id <- c('titanic_training')
getSessionId(model_id)
}

