% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{getSummary}
\alias{getSummary}
\title{Creates a data.frame with the mean,sd,median,st,max,min, and number of NAs in a given dataframe or ta.data.frames}
\usage{
getSummary(dataset)
}
\arguments{
\item{dataframe}{to summerize}
}
\value{
The summary
}
\description{
Creates a data.frame with the mean,sd,median,st,max,min, and number of NAs in a given dataframe or ta.data.frames
}
\examples{
train_tadf <- ta.data.frame("select * from model_factory.titanic_train",sourceType="query")
getSummary(train_tadf)

train_df <- ta.pull(train_tadf)
getSummary(train_df)
}

