% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{getValidColumns}
\alias{getValidColumns}
\title{Returns all columns from a given table that have been cleared to be used by models}
\usage{
getValidColumns(get.test.columns.also = FALSE)
}
\arguments{
\item{get.test.columns.also}{If set to TRUE columns that have test_column =1 will also be added}
}
\value{
The result of taQuery to execute the installed R script. Session_id is returned as result as well
}
\description{
Returns all columns from a given table that have been cleared to be used by models
}
\examples{
formula <- as.formula(paste("churn_indicator ~ " paste(getValidColumns(schema="model_factory.", table_name="bom_loc"),collapse=" + "),sep=""))
}

