% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{getVarimp}
\alias{getVarimp}
\title{Get variable importance from a given model}
\usage{
getVarimp(model_type, model)
}
\arguments{
\item{model_type}{Type of model ("glm","ta.glm","randomForest","ta.forest")}

\item{model}{Your model of type lm/glm/ta.glm/ta.forest/randomforest}
}
\value{
Dataframe with variable importance data
}
\description{
Get variable importance from a given model
}
\examples{
train_tadf <- ta.data.frame("select * from model_factory.titanic_train",sourceType="query")
train <- ta.pull(train_tadf)

formula_glm <- as.formula(survived_int ~ sex + pclass + age + family + embarked + title)
mod <- glm(formula_glm,train,family="binomial")
getVarimp(model_type="glm",mod)

formula <- as.formula(survived ~ sex + pclass + age + family + embarked + title)
mod <- randomForest(formula,data= train,ntree=100,mtry=4,keep.forest=TRUE)
getVarimp(model_type="randomForest",mod)

formula_glm <- as.formula(survived_int ~ sex + pclass + age + family + embarked + title)
mod <- ta.glm(formula= formula_glm, data = train_tadf,family=binomial())
getVarimp(model_type="ta.glm",mod)

formula <- as.formula(survived ~ sex + pclass + age + family + embarked + title)
mod <- ta.forest(formula= formula, data = train,type = "classification", ntree = 100,mtry=4)
getVarimp(model_type="ta.forest",mod)
}

