% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{installAndRunOnAster}
\alias{installAndRunOnAster}
\title{Installes the script on the Aster cluster before running it.
It will remove any previous script with the same name}
\usage{
installAndRunOnAster(path, script, mem = "unlimited")
}
\arguments{
\item{path}{The path to the script file with the trailing slash}

\item{script}{Script you want to run on Aster}

\item{mem}{The amount of memory that the R session is allowed ot use id MB, default is unlimited}
}
\value{
The result of taQuery to execute the installed R script. Session_id is returned as result as well
}
\description{
Installes the script on the Aster cluster before running it.
It will remove any previous script with the same name
}
\examples{
installAndRunOnAster('C:/projects/awesome_model/','save_kpn.R')
or
installAndRunOnAster('C:/projects/awesome_model/','save_kpn_with_less_memory.R','1000')
}

