% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pullAccuracy}
\alias{pullAccuracy}
\title{Pull Accuracy from model_factory.model_test_results table based on session_id, threshold_value and threshold_type}
\usage{
pullAccuracy(session_id, threshold_value, threshold_type)
}
\arguments{
\item{session_id}{Existing session_id}

\item{threshold_value}{Threshold_value}

\item{threshold_type}{Threshold type ("population" or "probability")}
}
\value{
Accuracy of existing session_id defined by given threshold_value and threshold_type
}
\description{
Pull Accuracy from model_factory.model_test_results table based on session_id, threshold_value and threshold_type
}
\examples{
pullAccuracy('titanic_training_vecht499_20160630_623433',0.5,"population")
pullAccuracy('titanic_training_vecht499_20160630_623433',0.5,"probability")
}

