% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pullBackTestCoordinates}
\alias{pullBackTestCoordinates}
\title{Pull backtesting data from model_factory.model_backtesting table based on session_id}
\usage{
pullBackTestCoordinates(session_id)
}
\arguments{
\item{session_id}{Vector of session id's}
}
\value{
Model backtest results of existing session_id
}
\description{
Pull backtesting data from model_factory.model_backtesting table based on session_id
}
\examples{
pullBackTestCoordinates(c('example_package_vecht499_20160630_351931'))
}

