% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pullConfMatrix}
\alias{pullConfMatrix}
\title{Pull Confusion matrix from model_factory.model_test_results table based on session_id, threshold_value and threshold_type}
\usage{
pullConfMatrix(session_id, threshold_value, threshold_type)
}
\arguments{
\item{session_id}{Existing session_id}

\item{threshold_value}{Threshold_value}

\item{threshold_type}{Threshold type ("population" or "probability")}
}
\value{
Confusion matrix of existing session_id defined by given threshold_value and threshold_type
}
\description{
Pull Confusion matrix from model_factory.model_test_results table based on session_id, threshold_value and threshold_type
}
\examples{
pullConfMatrix('titanic_training_vecht499_20160630_623433',0.5,"population")
pullConfMatrix('titanic_training_vecht499_20160630_623433',0.5,"probability")
}

