% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pullLiftChart}
\alias{pullLiftChart}
\title{Pull Lift chart from model_factory.model_test_results table based on session_id}
\usage{
pullLiftChart(session_id)
}
\arguments{
\item{session_id}{Vector of session id's}
}
\value{
Lift chart table of existing session_id
}
\description{
Pull Lift chart from model_factory.model_test_results table based on session_id
}
\examples{
pullLiftChart(c('titanic_training_vecht499_20160630_623433'))
}

