% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pullModel}
\alias{pullModel}
\title{Retrives a model object stored with storeModel() function in session with session ID session_id.}
\usage{
pullModel(session_id_in)
}
\arguments{
\item{session_id_in}{The session Id of the session that stored the model}
}
\value{
The result of the clean up
}
\description{
Retrives a model object stored with storeModel() function in session with session ID session_id.
}
\examples{
forest <- pullModel("ols_session_id")
varimp <- getVarimp("randomForest",forest)
forest_prediction <-  as.data.frame(predict(forest,newdata= test,"prob"))
}

