% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pullModelScores}
\alias{pullModelScores}
\title{Pull model scores from model_factory.model_scores table based on session_id}
\usage{
pullModelScores(session_id)
}
\arguments{
\item{session_id}{Vector of session id's}
}
\value{
Model scores of existing session_id
}
\description{
Pull model scores from model_factory.model_scores table based on session_id
}
\examples{
pullModelScores(c('titanic_training_vecht499_20160630_623433'))
}

