% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pullROC}
\alias{pullROC}
\title{Pull ROC from model_factory.model_test_results table based on session_id}
\usage{
pullROC(session_id)
}
\arguments{
\item{session_id}{Vector of session id's}
}
\value{
ROC table of existing session_id
}
\description{
Pull ROC from model_factory.model_test_results table based on session_id
}
\examples{
pullROC(c('titanic_training_vecht499_20160630_623433'))
}

