% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pullSummary}
\alias{pullSummary}
\title{Pull the summary of existing session_id from model_factory.model_summary table}
\usage{
pullSummary(session_id)
}
\arguments{
\item{session_id}{Vector of session id's}
}
\value{
Summary of existing session_id
}
\description{
Pull the summary of existing session_id from model_factory.model_summary table
}
\examples{
pullSummary(c('titanic_training_vecht499_20160629_105350'))
}

