% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pullTestResults}
\alias{pullTestResults}
\title{Pull test results from model_factory.model_test_results table based on session_id}
\usage{
pullTestResults(session_id)
}
\arguments{
\item{session_id}{Vector of session id's}
}
\value{
Test results of existing session_id
}
\description{
Pull test results from model_factory.model_test_results table based on session_id
}
\examples{
pullTestResults(c('titanic_training_vecht499_20160630_623433'))
}

