% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pullVarimp}
\alias{pullVarimp}
\title{Pull the variable importance of existing session_id from model_factory.model_varimp table}
\usage{
pullVarimp(session_id)
}
\arguments{
\item{session_id}{Vector of session id's}
}
\value{
Summary of existing session_id
}
\description{
Pull the variable importance of existing session_id from model_factory.model_varimp table
}
\examples{
pullVarimp(c('titanic_training_vecht499_20160630_364493'))
}

