% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{runOnAster}
\alias{runOnAster}
\title{Run script on Aster cluster}
\usage{
runOnAster(script, mem = "unlimited")
}
\arguments{
\item{script}{Script you want to run on Aster}

\item{mem}{The amount of memory that the R session is allowed ot use id MB, default is unlimited}
}
\value{
The result of taQuery to execute the installed R script. Session_id is returned as result as well
}
\description{
Run script on Aster cluster
}
\examples{
ta.install.scripts('C:/Users/vecht499/Desktop/test.R')
runOnAster('test.R')
runOnAster('test.R','1000')
}

