% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{storeBackTestCoordinates}
\alias{storeBackTestCoordinates}
\title{Store the backtesting data for a time series model.}
\usage{
storeBackTestCoordinates(predicted_value, actual_value, period)
}
\arguments{
\item{predicted_value}{The predictions}

\item{actual_value}{The the real value that the model was trying to predict}

\item{period}{The time axis (POSIXlt)}
}
\value{
The coordenates for a backtesting plot
}
\description{
Store the backtesting data for a time series model.
}
\examples{
model_id <- c('example_package')
addModelId(model_id,'Example for ModelFactoryR package','aggregated')
getSessionId(model_id)
df <- data.frame("predicted_value"=c(1.5,2.3,3.1,4.9,5.7),"actual_value"=c(1.2,3.0,3.7,4.3,5.1),
"period"= as.POSIXlt(c('2015-01-01','2015-01-02','2015-01-03','2015-01-04','2015-01-05'),"Europe/Amsterdam"))
storeBackTestCoordinates(df$predicted_value,df$actual_value,df$period)
}

