% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{storeModel}
\alias{storeModel}
\title{Stores a model object as rows in a the Aster Database.
The object must be serializeable, so vectors, lists, dataframes}
\usage{
storeModel(model)
}
\arguments{
\item{model}{The model object to store}
}
\value{
The result of the clean up
}
\description{
Stores a model object as rows in a the Aster Database.
The object must be serializeable, so vectors, lists, dataframes
}
\examples{
titanic_dataset <- taQuery("select * from modelfactory.titanic_dataset")
train <- titanic_dataset[titanic_dataset$train==1,]
formula <- as.formula(survived ~ sex + pclass + age + family + embarked + title)
forest <- randomForest(formula,data= train,ntree=100,mtry=4,keep.forest=TRUE)
storeModel(forest)
}

