% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{storeModelScores}
\alias{storeModelScores}
\title{Write the scores to model_factory.model_scores table}
\usage{
storeModelScores(id, scores)
}
\arguments{
\item{id}{Customer identifyer}

\item{scores}{Scores}
}
\value{
The result of taQuery to update the model_factory.model_scores table
}
\description{
Write the scores to model_factory.model_scores table
}
\examples{
model_id <- c('titanic_training')
getSessionId(model_id)
train <- ta.data.frame("select * from model_factory.titanic_train",sourceType="query")
test <- ta.data.frame("select * from model_factory.titanic_test",sourceType="query")
formula <- "survived ~ sex + pclass + age + family + embarked + title"
ta_forest <- ta.forest(formula= formula, data = train,type = "classification", ntree = 100,mtry=4)
ta_forest_prediction <- ta.forest.predict(object = ta_forest, newdata=test, id.col="passengerid")
forest_output <- ta_forest_prediction$retval
forest_output$pred <- ifelse(forest_output$prediction=='Survived',as.numeric(forest_output$confidence_upper),
                             1-as.numeric(forest_output$confidence_upper))
test <- ta.pull(test)
storeModelScores(test$passengerid,forest_output$pred)
}

