% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{storeTestResults}
\alias{storeTestResults}
\title{Write the test results to model_factory.model_test_results table}
\usage{
storeTestResults(dataset)
}
\arguments{
\item{dataset}{Dataframe which is the result of getTestResults to be added unto the summary table}
}
\value{
The result of taQuery to update the model_factory.model_test_results table
}
\description{
Write the test results to model_factory.model_test_results table
}
\examples{
model_id <- c('titanic_training')
getSessionId(model_id)
train <- ta.data.frame("select * from model_factory.titanic_train",sourceType="query")
test <- ta.data.frame("select * from model_factory.titanic_test",sourceType="query")

formula <- "survived ~ sex + pclass + age + family + embarked + title"
ta_forest <- ta.forest(formula= formula, data = train,type = "classification", ntree = 100,mtry=4)
ta_forest_prediction <- ta.forest.predict(object = ta_forest, newdata=test, id.col="passengerid")

forest_output <- ta_forest_prediction$retval
forest_output$pred <- ifelse(forest_output$prediction=='Survived',as.numeric(forest_output$confidence_upper),
                             1-as.numeric(forest_output$confidence_upper))

test <- ta.pull(test)
test_results <- getTestResults(forest_output$pred,test$survived_int)
storeTestResults(test_results)
}

