% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{storeVarImp}
\alias{storeVarImp}
\title{Store a dataframe/vectors filled with the variable importance of the model.
Standard deviation and coefficients are optional}
\usage{
storeVarImp(variable, importance, std, coefficients)
}
\arguments{
\item{variable}{The variables}

\item{importance}{The importance of the variables}

\item{std}{The standard diviation}

\item{coefficients}{The coefficients}

\item{session_id}{The current session ID}
}
\value{
The result of the insert taQuery
}
\description{
Store a dataframe/vectors filled with the variable importance of the model.
Standard deviation and coefficients are optional
}
\examples{
model_id <- c('titanic_training')
getSessionId(model_id)

train <- ta.data.frame("select * from model_factory.titanic_train",sourceType="query")
test <- ta.data.frame("select * from model_factory.titanic_test",sourceType="query")

formula <- "survived ~ sex + pclass + age + family + embarked + title"
ta_forest <- ta.forest(formula= formula, data = train,type = "classification", ntree = 100,mtry=4)
varimp <- getVarimp("ta.forest",ta_forest)
storeVarImp(varimp$variable,varimp$importance)
}

