% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{updateThreshold}
\alias{updateThreshold}
\title{Update threshold information in model_factory.model_overview table}
\usage{
updateThreshold(model_id, threshold_value, threshold_type)
}
\arguments{
\item{model_id}{The current model ID}

\item{threshold_value}{Chosen threshold_value}

\item{threshold_type}{Threshold type ("population" or "probability")}
}
\value{
The result of taQuery to update the model_factory.model_overview table
}
\description{
Update threshold information in model_factory.model_overview table
}
\examples{
updateThreshold('titanic_training','0.5','probability')
}

